+++ backend/src/neuroagent/app/dependencies.py
@@ -111,6 +111,7 @@ from neuroagent.tools import (
     StrainGetOneTool,
     SubjectGetAllTool,
     SubjectGetOneTool,
+    WeatherTool,
     WebSearchTool,
 )
 from neuroagent.tools.base_tool import BaseTool
@@ -467,7 +468,8 @@ def get_tool_list(
         RunPythonTool,
         WebSearchTool,
         # NowTool,
-        # WeatherTool,
+        WeatherTool,
+        # RandomPlotGeneratorTool,
     ]
 
     all_tools = internal_tool_list + mcp_tool_list