+++ CHANGELOG.md
@@ -33,6 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Exa tools no longer MCP.
+- Adapt everything to vercel v5.
+- Adapt to Response API.
+
 
 ## [v0.11.5] - 6.11.2025
 
@@ -90,6 +93,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use typescript autogen for backend types in frontend.
 - Try to enforce using metric tools rather than downloading assets.
 - Rule to avoid overvalidating.
+
 
 ## [v0.10.0] - 2.10.2025